Refactor Sheffield interface into service and api client, add tests
