+++ spec/fixtures/pv_live/0.json
@@ -1,0 +1,1 @@
+{"data":[[0,"2021-01-02T00:00:00Z",1,0.0,13080.0],[0,"2021-01-01T23:30:00Z",1,0.0,13080.0],[0,"2021-01-01T23:00:00Z",1,0.0,13080.0],[0,"2021-01-01T22:30:00Z",1,0.0,13080.0],[0,"2021-01-01T22:00:00Z",1,0.0,13080.0],[0,"2021-01-01T21:30:00Z",1,0.0,13080.0],[0,"2021-01-01T21:00:00Z",1,0.0,13080.0],[0,"2021-01-01T20:30:00Z",1,0.0,13080.0],[0,"2021-01-01T20:00:00Z",1,0.0,13080.0],[0,"2021-01-01T19:30:00Z",1,0.0,13080.0],[0,"2021-01-01T19:00:00Z",1,0.0,13080.0],[0,"2021-01-01T18:30:00Z",1,0.0,13080.0],[0,"2021-01-01T18:00:00Z",1,0.0,13080.0],[0,"2021-01-01T17:30:00Z",1,0.0,13080.0],[0,"2021-01-01T17:00:00Z",1,0.0,13080.0],[0,"2021-01-01T16:30:00Z",1,1.8,13080.0],[0,"2021-01-01T16:00:00Z",1,37.1,13080.0],[0,"2021-01-01T15:30:00Z",1,177.0,13080.0],[0,"2021-01-01T15:00:00Z",1,467.0,13080.0],[0,"2021-01-01T14:30:00Z",1,771.0,13080.0],[0,"2021-01-01T14:00:00Z",1,974.0,13080.0],[0,"2021-01-01T13:30:00Z",1,1060.0,13080.0],[0,"2021-01-01T13:00:00Z",1,1120.0,13080.0],[0,"2021-01-01T12:30:00Z",1,1080.0,13080.0],[0,"2021-01-01T12:00:00Z",1,1060.0,13080.0],[0,"2021-01-01T11:30:00Z",1,932.0,13080.0],[0,"2021-01-01T11:00:00Z",1,714.0,13080.0],[0,"2021-01-01T10:30:00Z",1,514.0,13080.0],[0,"2021-01-01T10:00:00Z",1,344.0,13080.0],[0,"2021-01-01T09:30:00Z",1,198.0,13080.0],[0,"2021-01-01T09:00:00Z",1,43.0,13080.0],[0,"2021-01-01T08:30:00Z",1,1.69,13080.0],[0,"2021-01-01T08:00:00Z",1,0.0,13080.0],[0,"2021-01-01T07:30:00Z",1,0.0,13080.0],[0,"2021-01-01T07:00:00Z",1,0.0,13080.0],[0,"2021-01-01T06:30:00Z",1,0.0,13080.0],[0,"2021-01-01T06:00:00Z",1,0.0,13080.0],[0,"2021-01-01T05:30:00Z",1,0.0,13080.0],[0,"2021-01-01T05:00:00Z",1,0.0,13080.0],[0,"2021-01-01T04:30:00Z",1,0.0,13080.0],[0,"2021-01-01T04:00:00Z",1,0.0,13080.0],[0,"2021-01-01T03:30:00Z",1,0.0,13080.0],[0,"2021-01-01T03:00:00Z",1,0.0,13080.0],[0,"2021-01-01T02:30:00Z",1,0.0,13080.0],[0,"2021-01-01T02:00:00Z",1,0.0,13080.0],[0,"2021-01-01T01:30:00Z",1,0.0,13080.0],[0,"2021-01-01T01:00:00Z",1,0.0,13080.0],[0,"2021-01-01T00:30:00Z",1,0.0,13080.0],[0,"2021-01-01T00:00:00Z",1,0.0,13080.0]],"meta":["gsp_id","datetime_gmt","n_ggds","generation_mw","installedcapacity_mwp"]}
